teckwah-TMS/main/requirements.txt
# 배송 실시간 관제 시스템 백엔드 요구사항
# Python 3.12.9 기준

# FastAPI 관련
fastapi==0.109.2
uvicorn==0.27.1
starlette==0.36.3

# 데이터베이스 관련
sqlalchemy==2.0.27
pymysql==1.1.0
cryptography==42.0.2

# 데이터 검증 및 직렬화
pydantic==2.6.1
pydantic-core==2.16.2
email-validator==2.1.0.post1

# 보안 관련
bcrypt==4.1.2
python-multipart==0.0.9
itsdangerous==2.1.2

# 템플릿 엔진 (서버 사이드 렌더링)
jinja2==3.1.3
aiofiles==23.2.1  # 비동기 파일 처리

# 환경 변수 관리
python-dotenv==1.0.1

# 로깅 및 유틸리티
colorama==0.4.6  # Windows 콘솔에서 색상 지원

# 날짜/시간 처리
pytz==2024.1

# 프로덕션 서버 (GAE용)
gunicorn==21.2.0

# 엑셀 파일 생성 
xlsxwriter==3.1.9
